requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosmtplib==4.0.2
alembic==1.16.5
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
priority==2.0.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
Quart==0.20.0
quart-cors==0.8.0
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1